backend/requirements.txt
Django==6.0.2
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
Pillow==11.1.0
djangorestframework-simplejwt==5.4.0
